Change floats to double
as per https://github.com/salilab/imp/pull/1042#discussion_r542826223

+++ modules/em/include/DensityMap.h
@@ -387,7 +387,7 @@ class IMPEMEXPORT DensityMap : public IMP::Object {
       \param[in] inverse Set to true to crop the particle volume of the map
       \return the new cropped map.
    */
-  DensityMap *get_cropped(Particles ps, float distance, bool inverse = false);
+  DensityMap *get_cropped(Particles ps, double distance, bool inverse = false);
 
   //! Create and return a new cropped map with the given bounding box extent
   /** \param[in] bb the bounding box
